requirements.txt
asgiref==3.8.1
Django==5.2.3
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
environs==14.2.0
drf-yasg==1.21.10